Project4_TestPlan_AIAgents/backend/pyproject.toml
[project]
name = "intelligent-test-plan-agent"
version = "0.1.0"
description = "FastAPI Backend for Intelligent Test Plan Generator"

[tool.poetry]
name = "intelligent-test-plan-agent"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"
groq = "^0.4.0"
pydantic = "^2.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
